feat: support PyPI distribution (`pip install stackchan-mcp`)
Make the gateway installable from PyPI as `stackchan-mcp`, with a
hardened publish pipeline.

Packaging
- Add a console_script entry point at `stackchan_mcp.cli:main` so
  `pipx install stackchan-mcp` exposes a `stackchan-mcp` CLI command.
- Move side effects (`load_dotenv()`, `logging.basicConfig()`) into
  `cli.main` so that `import stackchan_mcp` is now side-effect-free.
- Add `gateway/LICENSE` and `license-files = ["LICENSE"]` so the
  published wheel/sdist include the MIT license text (PEP 639).

Publish workflow (.github/workflows/publish.yml)
- Tag-based (`v*`) build + Trusted Publishing.
- Refuses to publish unless the tag commit is an ancestor of
  `origin/main`, so feature-branch tags cannot push to PyPI.
- Verifies the tag matches `gateway/pyproject.toml` after PEP 440
  normalization. Requires a `v` prefix and rejects local versions;
  pre-releases (rc/a/b) are allowed.
- Runs `uv run ruff check .` and `uv run pytest` before building.
- Pins `pypa/gh-action-pypi-publish` to a specific minor.patch tag,
  uses major-version pins for `actions/checkout` and
  `astral-sh/setup-uv`, and uses top-level `permissions: {}` with
  per-job opt-in.

Docs
- README.md / README.ja.md: recommend `uv tool install` / `pipx install`
  for end users; keep the source-tree uv path for contributors.
- gateway/README.md: add an Install (end users) section (this is the
  PyPI long_description), a license-boundary note for the GPL/MIT
  split, and a note that the package name, CLI command, and MCP
  server id are all `stackchan-mcp`.
- CONTRIBUTING.md: document release gates, one-time setup (Trusted
  Publishing, environment, protected tags), per-release steps, the
  yank policy, and the workflow's pinning policy.

Operational follow-ups (CHANGELOG, fork-friendly token path) are
tracked in #45.

Closes #11
